Organize simple tree code
