+++ packages/dds/tree/src/simple-tree/api/getSimpleSchema.ts
@@ -3,9 +3,9 @@
  * Licensed under the MIT License.
  */
 
-import { getOrCreate } from "../util/index.js";
-import type { TreeNodeSchema } from "./core/index.js";
-import type { ImplicitAllowedTypes } from "./schemaTypes.js";
+import { getOrCreate } from "../../util/index.js";
+import type { TreeNodeSchema } from "../core/index.js";
+import type { ImplicitAllowedTypes } from "../schemaTypes.js";
 import type { SimpleTreeSchema } from "./simpleSchema.js";
 import { toSimpleTreeSchema } from "./viewSchemaToSimpleSchema.js";
