+++ packages/dds/tree/src/simple-tree/api/jsonSchema.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import type { NodeKind } from "./core/index.js";
+import type { NodeKind } from "../core/index.js";
 
 /**
  * The fully-qualified {@link TreeNodeSchemaCore.identifier}.
